Refactor bug
